requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
numpy==1.26.2
Pillow==10.1.0
scikit-learn==1.3.2
scikit-image==0.22.0
python-multipart==0.0.6
